refactor(App): move availableTimes state into ReservationForm.js
