+++ src/App.js
@@ -6,30 +6,14 @@ import MainPage from './pages/MainPage/MainPage';
 import ReservationPage from './pages/ReservationPage/ReservationPage';
 import ConfirmationPage from './pages/ConfirmationPage/ConfirmationPage';
 import { BrowserRouter as Router, Routes, Route} from 'react-router-dom';
-import { useReducer } from 'react';
-
-export function initializeTimes() {
-	const today = new Date();
-	return window.fetchAPI(today);
-}
-
-export function updateTimes(state, action) {
-	if (action.type === "update") {
-		const date = new Date(action.date);
-		return window.fetchAPI(date);
-	}
-	return state;
-}
 
 function App() {
-	const [availableTimes, dispatch] = useReducer(updateTimes, initializeTimes());
-
 	return (
 		<Router>
 			<Header />
 			<Routes>
 				<Route path="/" element={<MainPage />} />
-				<Route path="/reservation" element={<ReservationPage availableTimes={availableTimes} dispatch={dispatch} />} />
+				<Route path="/reservation" element={<ReservationPage />} />
         		<Route path="/confirmation" element={<ConfirmationPage />} />
 			</Routes>
 			<Footer />
